+++ CHANGELOG.md
@@ -29,16 +29,19 @@ in 3.17. Use `PaymentSettings.defaultTransactionFlowStrategy` instead.
 - Add `NOTIFY_CUSTOMER` flag to `FulfillmentApproved` type - #13637, by @Air-t
   - Inform apps if customer should be notified when fulfillment is approved.
 - Add `GIFT_CARD_EXPORT_COMPLETED` webhook - #13765, by @Air-t
-  - Event sent when CSV export is completed.
+  - Event sent when CSV export for gift cards is completed.
+- Add `PRODUCT_EXPORT_COMPLETED` webhook - #13787, by @Air-t
+  - Event sent when CSV export for products is completed.
 
 - Add `FULFILLMENT_TRACKING_NUMBER_UPDATED` webhook - #13708, by @Air-t
   - Called after `fulfillmentUpdateTracking` or `orderFulfill` mutation if tracking number is updated.
 
 ### Other changes
 - Fix error in variant available stock calculation - 13593 by @awaisdar001
+- Add missing currency in action for transaction requests - 13786 by @AjmalPonneth
 - Fix giftcard code description - #13728 by @rafiwts
-
 - Change error message when denying a permission - #13334 by @rafiwts
+- Add filter by slugs to attribute choices - #13761 by @rafiwts
 
 # 3.15.0 [Unreleased]